+++ src/main/java/codesquad/WebServer.java
@@ -3,8 +3,8 @@ package codesquad;
 import codesquad.config.GlobalConfig;
 import codesquad.http.HttpProcessor;
 import codesquad.http.HttpRequestProcessor;
-import codesquad.http.RequestHandlerResolver;
-import codesquad.http.StaticResourceRequestHandler;
+import codesquad.http.handler.RequestHandlerResolver;
+import codesquad.http.handler.StaticResourceRequestHandler;
 import codesquad.http.parser.HttpParser;
 import codesquad.socket.ClientSocket;
 import codesquad.socket.ServerSocket;
