+++ src/main/java/codesquad/http/handler/RequestHandlerResolver.java
@@ -1,4 +1,4 @@
-package codesquad.http;
+package codesquad.http.handler;
 
 import java.util.Map;
